Demonstration: Scala plugin provider as plugin itself
This is only a show case that the plugin provider for different JVM
language (Scala in this case) can be a plugin itself.

This change depends on core change [1] with this modification [2].
As example this plugin can be used [3]. To make it work: patch core,
compile plugin-api, compile cookbook-plugin with this change, put these
dependent libraries under `$gerrit_site/lib` directory:

* scala-compiler-2.10.3.jar
* scala-library-2.10.3.jar
* scala-reflect-2.10.3.jar

And drop the cookbook-plugin & review.scala plugins in
`$gerrit_site/plugins`.

ssh gerrit review --help should produce:

Available commands of review are:

   approve
   dislike
   recommend
   reject

See 'review COMMAND --help' for more information.

[1] https://gerrit-review.googlesource.com/54275
[2] https://gist.github.com/davido/8891691
[3] https://github.com/davido/gerrit-scala-plugin/blob/master/review.scala

Change-Id: I1c8f20d5bf1e182c07b41009d6ef98f7575449dd