Demo of Issue 2737(prolog rule reload from plugin)
The predicate is defined in pgk_one package.
It is defined in class: PRED_example_1.java
So to see the problem put rule.pl in refs/meta/config
with the following content:
submit_rule(submit(X)):-pkg_one:example(X).

After first load of the rule it gets initialized
and it does not consult the plugin anymore.
So if one removes the plugin the predicate is still
visible to prolog. That means no possibility to update.
The only way to update the predicate is to deploy
new version of plugin and restart gerrit.

Change-Id: Ia64053639883beec25fc13d9fbe10d6e0a7d64e7
Signed-off-by: Eryk Szymanski <eryksz@gmail.com>