Fix documentation: wrong invocation of plugin.
- The synopsis for the cookbook print ssh command was wrong.
- The examples provided for calling the ssh api for the
commands "print" and "admin" was wrong.

In all cases described above the plugin was referenced as
"cookbook-plugin" instead of "cookbook" in the ssh invocation.

Now "cookbook" is used for referencing the plugin consistently
in the synopsis and in the examples.

Up to now a variable @PLUGIN@ was used in the sources, but the
value of this variable is obviously "cookbook-plugin" which
is not the appropriate value here. May be there is a better
approach than hard coding "cookbook" here.

Change-Id: I3e44c4397944f17dec17098392d562f124d1cfe1