[DO NOT SUBMIT] Show that onFailure() in plugin is never invoked
When a plugin invokes a REST endpoint in the UI, it provides an
`AsyncCallback` to handle the result. At the moment the
`onFailure(Throwable)` of the callback is never invoked, even if there
is an error. Errors are always handled by the Gerrit core UI which
shows the error dialog. This means currently plugins cannot do any
error handling and e.g. ignore expected errors.

This change demonstrates this behaviour. When the index screen of the
cookbook plugin is opened, it sends a REST request to the server which
results in '404 Not Found'. Although the onFailure() method is empty,
which should ignore any failure, an error dialog is displayed. This is
because the error is handled by Gerrit core which shows the error
dialog.

Change-Id: I4fbca85122a4252ae235442d64a6d50bec21f929
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>