Update patch set 9
Patch Set 9: Verified-1

Only tree (non-standalone) mode can be verified successfully for now, as there is an issue with standalone: trying the latter leads to a security exception for us. That exception in the debugger shows some problem with the way junit is now repackaged into the plugin jar, or so it seems to me. The exception is a java.lang.SecurityException: Invalid signature file digest for Manifest main attributes.

Patch-set: 9